Use user-provided directories in `gem server`
Previously Gem::Server would reset the gem list for certain requests.
This was designed to always provide the latest installed gems for
installation.

When the --dir option was used, resetting the gem list would also revert
to the standard gem directories.  This meant the user-provided
directories would always be ignored.

Now the reset includes the gem directories the user provided to
eliminate this problem.

Fixes #696